requirements-hash.txt
# This file is autogenerated by pip-compile with python 3.10
pretty-py3==0.2.4 \
    --hash=sha256:0953e0f475b2a76a89e06d643dad11509f771736c7c6aacd1efb9ae50b27f564 \
    --hash=sha256:92d0d3eb75051ff0aa0dfa72bfdf36c73fb1b54ce7ee9400645c3a9f18dcd16a
